+++ src/sklweka/classifiers.py
@@ -62,8 +62,8 @@ class WekaEstimator(BaseEstimator, OptionHandler, RegressorMixin, ClassifierMixi
         self.header_ = None
         self.classes_ = None
         # the following references are required for get_params/set_params
-        self._classname = classname
-        self._options = options
+        self._classname = classname if (classname is not None) else self._classifier.classname
+        self._options = options if (options is not None) else self._classifier.options
         self._nominal_input_vars = nominal_input_vars
         self._nominal_output_var = nominal_output_var
         self._num_nominal_input_labels = num_nominal_input_labels