+++ CHANGES.rst
@@ -8,6 +8,8 @@ Changelog
   distinguishes between classification and regression to return the correct score.
 - renamed `data` to `X` and `targets` to `y`, since some sklearn schemes use named arguments
 - added dummy argument `sample_weight=None` to `fit`, `score` and `fit_predict` methods
+- fixed: when supplying Classifier or JBObject instead of classname/options, classname/options
+  now get determined automatically
 
 
 0.0.6 (2022-04-26)